+++ src/entities/vessel/index.ts
@@ -1,2 +1,2 @@
-export type { LatLon } from './vessel.svelte';
+export type { LatLon } from '$shared/signalk';
 export { OwnVessel } from './vessel.svelte';